chore : #2 각 화면에 따른 파일 분리
